# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ramune Catcher API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Flavor {
  id: number;
  /** Japanese name (primary) */
  japaneseName: string;
  /** English name (secondary) */
  name: string;
  barcode: string;
  /** Hex color representing the flavor */
  color: string;
  /** Brand name e.g. Hata, Doraemon, Sangaria */
  brand: string;
  /** Category e.g. standard, limited, savory, doraemon */
  category: string;
  sortOrder: number;
  description?: string;
}

export interface CaughtFlavor {
  id: number;
  flavorId: number;
  caughtAt: string;
}

export interface CatchFlavorBody {
  flavorId: number;
}

export interface Location {
  id: number;
  name: string;
  city: string;
  country: string;
  lat: number;
  lng: number;
  confirmedCount: number;
  addedBy?: string;
}

export interface LocationFlavorEntry {
  flavor: Flavor;
  /** Price in local currency */
  price?: number;
  /** Currency code e.g. SEK, JPY, USD */
  currency?: string;
  /** Username who confirmed this */
  addedBy?: string;
  addedAt: string;
}

export interface LocationWithFlavors {
  id: number;
  name: string;
  city: string;
  country: string;
  lat: number;
  lng: number;
  confirmedCount: number;
  addedBy?: string;
  flavors: LocationFlavorEntry[];
}

export interface LocationFlavor {
  id: number;
  locationId: number;
  flavorId: number;
  price?: number;
  currency?: string;
  addedBy?: string;
  addedAt: string;
}

export interface CreateLocationBody {
  name: string;
  city: string;
  country: string;
  lat: number;
  lng: number;
  addedBy?: string;
}

export interface AddLocationFlavorBody {
  flavorId: number;
  price?: number;
  currency?: string;
  addedBy?: string;
}

export interface RemoveLocationFlavorBody {
  flavorId: number;
}

export interface Stats {
  totalFlavors: number;
  caughtFlavors: number;
  totalLocations: number;
  recentlyCaught: CaughtFlavor[];
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  createdAt: string;
}

export interface CreateUserBody {
  username: string;
  displayName: string;
}

export interface AddFriendBody {
  friendUsername: string;
}