# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ramune Catcher API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: flavors
  - name: caught
  - name: locations
  - name: stats
  - name: users
  - name: friends
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /flavors:
    get:
      operationId: listFlavors
      tags: [flavors]
      summary: List all ramune flavors
      responses:
        "200":
          description: List of flavors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flavor"

  /flavors/{id}:
    get:
      operationId: getFlavor
      tags: [flavors]
      summary: Get a single flavor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The flavor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flavor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /flavors/barcode/{barcode}:
    get:
      operationId: getFlavorByBarcode
      tags: [flavors]
      summary: Look up a flavor by barcode
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The flavor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flavor"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /caught:
    get:
      operationId: listCaught
      tags: [caught]
      summary: List all caught flavor IDs
      responses:
        "200":
          description: List of caught records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaughtFlavor"
    post:
      operationId: catchFlavor
      tags: [caught]
      summary: Mark a flavor as caught
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatchFlavorBody"
      responses:
        "201":
          description: Caught successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaughtFlavor"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already caught
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /caught/{flavorId}:
    delete:
      operationId: uncatchFlavor
      tags: [caught]
      summary: Remove a caught flavor
      parameters:
        - name: flavorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List all snack map locations
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Add a new snack location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /locations/{id}:
    get:
      operationId: getLocation
      tags: [locations]
      summary: Get a location with its confirmed flavors
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationWithFlavors"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateLocation
      tags: [locations]
      summary: Update a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /locations/{id}/flavors:
    post:
      operationId: addLocationFlavor
      tags: [locations]
      summary: Add a confirmed flavor to a location with optional price
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddLocationFlavorBody"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationFlavor"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: removeLocationFlavor
      tags: [locations]
      summary: Remove a confirmed flavor from a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoveLocationFlavorBody"
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get overall collection stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create or get a user by username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "200":
          description: User created or retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{username}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{username}/friends:
    get:
      operationId: getFriends
      tags: [friends]
      summary: Get friends list for a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of friends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: addFriend
      tags: [friends]
      summary: Add a friend
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFriendBody"
      responses:
        "201":
          description: Friend added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already friends
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{username}/friends/{friendUsername}:
    delete:
      operationId: removeFriend
      tags: [friends]
      summary: Remove a friend
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: friendUsername
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Flavor:
      type: object
      properties:
        id:
          type: integer
        japaneseName:
          type: string
          description: Japanese name (primary)
        name:
          type: string
          description: English name (secondary)
        barcode:
          type: string
        color:
          type: string
          description: Hex color representing the flavor
        brand:
          type: string
          description: Brand name e.g. Hata, Doraemon, Sangaria
        category:
          type: string
          description: Category e.g. standard, limited, savory, doraemon
        sortOrder:
          type: integer
        description:
          type: string
      required:
        - id
        - japaneseName
        - name
        - barcode
        - color
        - brand
        - category
        - sortOrder

    CaughtFlavor:
      type: object
      properties:
        id:
          type: integer
        flavorId:
          type: integer
        caughtAt:
          type: string
          format: date-time
      required:
        - id
        - flavorId
        - caughtAt

    CatchFlavorBody:
      type: object
      properties:
        flavorId:
          type: integer
      required:
        - flavorId

    Location:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        country:
          type: string
        lat:
          type: number
        lng:
          type: number
        confirmedCount:
          type: integer
        addedBy:
          type: string
      required:
        - id
        - name
        - city
        - country
        - lat
        - lng
        - confirmedCount

    LocationWithFlavors:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        country:
          type: string
        lat:
          type: number
        lng:
          type: number
        confirmedCount:
          type: integer
        addedBy:
          type: string
        flavors:
          type: array
          items:
            $ref: "#/components/schemas/LocationFlavorEntry"
      required:
        - id
        - name
        - city
        - country
        - lat
        - lng
        - confirmedCount
        - flavors

    LocationFlavorEntry:
      type: object
      properties:
        flavor:
          $ref: "#/components/schemas/Flavor"
        price:
          type: number
          description: Price in local currency
        currency:
          type: string
          description: Currency code e.g. SEK, JPY, USD
        addedBy:
          type: string
          description: Username who confirmed this
        addedAt:
          type: string
          format: date-time
      required:
        - flavor
        - addedAt

    LocationFlavor:
      type: object
      properties:
        id:
          type: integer
        locationId:
          type: integer
        flavorId:
          type: integer
        price:
          type: number
        currency:
          type: string
        addedBy:
          type: string
        addedAt:
          type: string
          format: date-time
      required:
        - id
        - locationId
        - flavorId
        - addedAt

    CreateLocationBody:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        country:
          type: string
        lat:
          type: number
        lng:
          type: number
        addedBy:
          type: string
      required:
        - name
        - city
        - country
        - lat
        - lng

    AddLocationFlavorBody:
      type: object
      properties:
        flavorId:
          type: integer
        price:
          type: number
        currency:
          type: string
        addedBy:
          type: string
      required:
        - flavorId

    RemoveLocationFlavorBody:
      type: object
      properties:
        flavorId:
          type: integer
      required:
        - flavorId

    Stats:
      type: object
      properties:
        totalFlavors:
          type: integer
        caughtFlavors:
          type: integer
        totalLocations:
          type: integer
        recentlyCaught:
          type: array
          items:
            $ref: "#/components/schemas/CaughtFlavor"
      required:
        - totalFlavors
        - caughtFlavors
        - totalLocations
        - recentlyCaught

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - displayName
        - createdAt

    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
      required:
        - username
        - displayName

    AddFriendBody:
      type: object
      properties:
        friendUsername:
          type: string
      required:
        - friendUsername